--- sftp-source-downloaders/sftp-source-downloader-cf-volume-services/src/main/java/io/spring/example/sftp/downloader/nfs/VolumeMount.java
@@ -1,57 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.spring.example.sftp.downloader.nfs;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * @author David Turanski
- **/
-class VolumeMount {
-	@JsonProperty("container_dir")
-	private String containerDir;
-
-	@JsonProperty("device_type")
-	private String deviceType;
-
-	@JsonProperty
-	private String mode;
-
-	public String getContainerDir() {
-		return containerDir;
-	}
-
-	public void setContainerDir(String containerDir) {
-		this.containerDir = containerDir;
-	}
-
-	public String getDeviceType() {
-		return deviceType;
-	}
-
-	public void setDeviceType(String deviceType) {
-		this.deviceType = deviceType;
-	}
-
-	public String getMode() {
-		return mode;
-	}
-
-	public void setMode(String mode) {
-		this.mode = mode;
-	}
-}
