+++ sftp-source-downloaders/sftp-source-downloader-cf-volume-services/src/main/java/org/springframework/cloud/stream/app/sftp/source/downloader/cf/volume/services/NFS.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.stream.app.sftp.source.downloader.cf.volume.services;
+
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.util.List;
+
+import org.springframework.util.CollectionUtils;
+
+/**
+ * @author David Turanski
+ **/
+abstract class NFS {
+
+	static VcapService load(InputStream json, String serviceName) {
+		List<VcapService> nfsService =  VcapServices.load(json).get(serviceName);
+		if (CollectionUtils.isEmpty(nfsService)) {
+			throw new RuntimeException(String.format("This application is not bound to service '%s'", serviceName));
+		}
+		if (nfsService.size() > 1) {
+			throw new RuntimeException("Expecting a unique nfs service, found " + nfsService.size());
+		}
+		return nfsService.get(0);
+	}
+
+	static VcapService load(String json, String serviceName) {
+		return load(new ByteArrayInputStream(json.getBytes()),
+			serviceName);
+	}
+
+}
