+++ sftp-source-downloaders/sftp-source-downloader-cf-volume-services/src/main/java/org/springframework/cloud/stream/app/sftp/source/downloader/cf/volume/services/NFSInputStreamPersister.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.stream.app.sftp.source.downloader.cf.volume.services;
+
+import java.io.File;
+import java.nio.file.Paths;
+
+import org.springframework.cloud.stream.app.sftp.source.downloader.core.AbstractFileInputStreamPersister;
+import org.springframework.cloud.stream.app.sftp.source.downloader.core.InputStreamTransfer;
+
+/**
+ * @author David Turanski
+ **/
+public class NFSInputStreamPersister extends AbstractFileInputStreamPersister {
+
+	final private String rootPath;
+
+	public NFSInputStreamPersister(String rootPath) {
+		this.rootPath = rootPath;
+	}
+
+	@Override
+	public void save(InputStreamTransfer transfer) {
+		File targetFile = Paths.get(rootPath, transfer.getTarget()).toFile();
+		doSave(transfer.getSource(), targetFile);
+	}
+}
