+++ sftp-source-downloaders/sftp-source-downloader-cf-volume-services/src/test/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
-org.springframework.boot.env.EnvironmentPostProcessor=io.spring.example.sftp.downloader.nfs\
-  .test.SftpDownloaderNFSTests$TestEnvironment
+org.springframework.boot.env.EnvironmentPostProcessor=org.springframework.cloud.stream.app.sftp.source.downloader\
+  .test.cf.volume.services.SftpDownloaderNFSTests$TestEnvironment
