+++ sftp-source-downloaders/sftp-source-downloader-local/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.spring.example.sftp.downloader.local.LocalInputStreamPersisterAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.springframework.cloud.stream.app.sftp.source.downloader.local.LocalInputStreamPersisterAutoConfiguration
