--- sftp-source-downloaders/sftp-source-downloader-s3/src/main/java/io/spring/example/sftp/downloader/s3/AmazonS3AutoConfiguration.java
@@ -1,61 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.spring.example.sftp.downloader.s3;
-
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.client.builder.AwsClientBuilder;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-/**
- * @author David Turanski
- **/
-@Configuration
-@EnableConfigurationProperties(AmazonS3ConfigurationProperties.class)
-public class AmazonS3AutoConfiguration {
-
-	@Configuration
-	@Profile("!cloud")
-	static class StandardAmazonS3AutoConfiguration {
-		@Bean
-		public AmazonS3 amazonS3(AmazonS3ConfigurationProperties properties) {
-
-			AWSCredentials awsCredentials = new BasicAWSCredentials(properties.getAccessKey(), properties.getSecretKey());
-
-			ClientConfiguration clientConfiguration = new ClientConfiguration();
-			clientConfiguration.setSignerOverride("AWSS3V4SignerType");
-
-			return AmazonS3ClientBuilder.standard()
-				.withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(properties.getEndpoint().toString(),
-					properties.getRegion().name()))
-				.withClientConfiguration(clientConfiguration)
-				.withPathStyleAccessEnabled(true)
-				.withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
-				.build();
-		}
-	}
-
-}
-
