+++ sftp-source-downloaders/sftp-source-downloader-s3/src/main/resources/META-INF/spring-configuration-metadata-whitelist.properties
@@ -1,2 +1,2 @@
-configuration-properties.classes=io.spring.example.sftp.downloader.s3.AmazonS3ConfigurationProperties
+configuration-properties.classes=org.springframework.cloud.stream.app.sftp.source.downloader.s3.AmazonS3ConfigurationProperties
