+++ sftp-source-downloaders/sftp-source-downloader-s3/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  io.spring.example.sftp.downloader.s3.S3InputStreamPersisterAutoConfiguration,\
-    io.spring.example.sftp.downloader.s3.AmazonS3AutoConfiguration
+  org.springframework.cloud.stream.app.sftp.source.downloader.s3.S3InputStreamPersisterAutoConfiguration,\
+    org.springframework.cloud.stream.app.sftp.source.downloader.s3.AmazonS3AutoConfiguration
